requirements.txt
Django>=3.1,<3.2
flake8>=3.8.4,<3.9.0
djangorestframework>=3.12.1,<3.13
psycopg2>=2.8.6,<2.9
django-rest-auth>=0.9.5,<0.10
django-rest-auth[with_social]
django-allauth
djangorestframework-jwt
